Write basic docs for crc_ functions.
Adding a baseline of docs for the generic crc_* functions with
explanations about their usage. Some of this was done in the README
already, but wanted information in the module as well so it get
published to the docs and is available in iex and editors.

The docs added are definitely not the best and mostly a 1st draft, but
something is better then nothing.