Update `crc.erl` to use crc_fast
Removed usages of legacy model specific CRC NIFs, except for crc_8.

For CRC-8 the generic model is not matching the legacy NIF, this needs
to be resolved at a later time so that it too can be removed.

This is the 1st step in reducing the code footprint for this library,
and hopefully on the road to a v1.0 release.