Adding `calculate/2`
Adding the `calculate/2` function to compliment the current `crc/2`
function but with the parameters reversed. This will allow piping binary
input from another function and passing parameters as the second
argument instead of the first.

Added `@moduledoc false` to legacy module so it doesn't appear in docs.

Update readme and app files for release of `v0.9`