Fix OTP 23 build by removing erl_interface refs
OTP 23 removed the `erl_interface` library so linker references to it
produce errors now. Luckily `erl_interface` isn't being used, so
deleting it fixes the problem.

This also deletes the reference to the `ei` library since it isn't being
used either. `ei` still exists, though.

Tested on OTP 22 and 23 on OSX and Linux.